refactor: move auth and gateway to internal
